Chore: 사용자 업데이트 요청 DTO 에서 선호재료, 기피재료 타입 List<Long> 으로 변경

+++ costcook/src/main/java/com/costcook/domain/request/UserUpdateRequest.java
@@ -12,9 +12,9 @@ public class UserUpdateRequest {
 
     private MultipartFile profileImage; // 선택
 
-    private List<Integer> preferences; // 필수
+    private List<Long> preferences; // 필수
 
-    private List<Integer> dislikedIngredients; // 필수
+    private List<Long> dislikedIngredients; // 필수
 
     private Boolean personalInfoAgreement; // 필수
 }
